* Refactored handlers (moved as child class) * removed GetAllTranslations query
