+++ docs/migrate-to-v7.md
@@ -20,6 +20,7 @@ Moved `Translate(this Enum target, params object[] formatArguments)` to `Localiz
 * introduced `IResourceRepository` - for easier storage implementations
 * moved `LocalizedResource` & `LocalizedTranslation` to `DbLocalizationProvider.Abstractions` project
 * moved `DiscoveredResource` from `DbLocalizationProvider.Sync` to `DbLocalizationProvider.Abstractions`
+* removed `GetAllTranslations` query
 
 
 ## Configuration
