+++ src/DbLocalizationProvider/Commands/CreateNewResources.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using DbLocalizationProvider.Abstractions;
 
 namespace DbLocalizationProvider.Commands
@@ -17,6 +18,64 @@ namespace DbLocalizationProvider.Commands
         /// </summary>
         /// <param name="e">Arguments obviously to understand what has been created.</param>
         public delegate void EventHandler(EventArgs e);
+
+        /// <summary>
+        /// Implementation of the command to create new resources
+        /// </summary>
+        public class Handler : ICommandHandler<Command>
+        {
+            private readonly ConfigurationContext _configurationContext;
+            private readonly IResourceRepository _repository;
+
+            /// <summary>
+            /// Creates new instance of the class.
+            /// </summary>
+            /// <param name="configurationContext">Configuration settings.</param>
+            /// <param name="repository">Resource repository</param>
+            public Handler(ConfigurationContext configurationContext, IResourceRepository repository)
+            {
+                _configurationContext = configurationContext;
+                _repository = repository;
+            }
+
+            /// <summary>
+            /// Handles the command. Actual instance of the command being executed is passed-in as argument
+            /// </summary>
+            /// <param name="command">Actual command instance being executed</param>
+            /// <exception cref="InvalidOperationException">Resource with key `{resource.ResourceKey}` already exists</exception>
+            public void Execute(Command command)
+            {
+                if (command.LocalizationResources == null || !command.LocalizationResources.Any())
+                {
+                    return;
+                }
+
+                foreach (var resource in command.LocalizationResources)
+                {
+                    var existingResource = _repository.GetByKey(resource.ResourceKey);
+
+                    if (existingResource != null)
+                    {
+                        throw new InvalidOperationException($"Resource with key `{resource.ResourceKey}` already exists");
+                    }
+
+                    resource.ModificationDate = DateTime.UtcNow;
+
+                    // if we are importing single translation and it's not invariant
+                    // set it also as invariant translation
+                    if (resource.Translations.Count == 1 && resource.Translations.InvariantTranslation() == null)
+                    {
+                        var t = resource.Translations.First();
+                        resource.Translations.Add(
+                            new LocalizationResourceTranslation { Value = t.Value, Language = string.Empty });
+                    }
+
+                    _repository.InsertResource(resource);
+
+                    _configurationContext.BaseCacheManager.StoreKnownKey(resource.ResourceKey);
+                }
+            }
+        }
 
         /// <summary>
         /// This command is usually used when creating new resources either from AdminUI or during import process
