+++ src/DbLocalizationProvider/Commands/RemoveTranslation.cs
@@ -3,7 +3,9 @@
 
 using System;
 using System.Globalization;
+using System.Linq;
 using DbLocalizationProvider.Abstractions;
+using DbLocalizationProvider.Cache;
 
 namespace DbLocalizationProvider.Commands
 {
@@ -12,6 +14,55 @@ namespace DbLocalizationProvider.Commands
     /// </summary>
     public class RemoveTranslation
     {
+        /// <summary>
+        /// Removes single translation
+        /// </summary>
+        public class Handler : ICommandHandler<Command>
+        {
+            private readonly ConfigurationContext _configurationContext;
+            private readonly IResourceRepository _repository;
+
+            /// <summary>
+            /// Creates new instance of the class.
+            /// </summary>
+            /// <param name="configurationContext">Configuration settings.</param>
+            /// <param name="repository">Resource repository</param>
+            public Handler(ConfigurationContext configurationContext, IResourceRepository repository)
+            {
+                _configurationContext = configurationContext;
+                _repository = repository;
+            }
+
+            /// <summary>
+            /// Handles the command. Actual instance of the command being executed is passed-in as argument
+            /// </summary>
+            /// <param name="command">Actual command instance being executed</param>
+            /// <exception cref="InvalidOperationException">Cannot delete translation for not modified resource (key: `{command.Key}`</exception>
+            public void Execute(Command command)
+            {
+                var resource = _repository.GetByKey(command.Key);
+
+                if (resource == null)
+                {
+                    return;
+                }
+
+                if (!resource.IsModified.HasValue || !resource.IsModified.Value)
+                {
+                    throw new InvalidOperationException(
+                        $"Cannot delete translation for not modified resource (key: `{command.Key}`");
+                }
+
+                var t = resource.Translations.FirstOrDefault(_ => _.Language == command.Language.Name);
+                if (t != null)
+                {
+                    _repository.DeleteTranslation(resource, t);
+                }
+
+                _configurationContext.CacheManager.Remove(CacheKeyHelper.BuildKey(command.Key));
+            }
+        }
+
         /// <summary>
         /// Execute this command if you need to just translation in some language for given resource.
         /// </summary>
